requirements.txt
ajaxuploader
asgi_redis
django-blockstack-auth
django-bootstrap3
django-datetime-widget
django-markdown-deux
django-rq
django-taggit
Django<2.0
dropbox
git+https://github.com/blockstack/blockstack-auth-python.git
gunicorn
mutagen
Pillow
psycopg2
pytz
requests>=2.5.1
rq
lxml
beautifulsoup4
